Run unit tests in parallel
